+++ src/Jacinda/Backend/TreeWalk.hs
@@ -37,7 +37,7 @@ data StreamError = NakedField
 
 instance Exception StreamError where
 
-(!) :: V.Vector a -> Int -> a
+(!) :: Show a => V.Vector a -> Int -> a
 v ! ix = case v V.!? ix of
     Just x  -> x
     Nothing -> throw $ IndexOutOfBounds ix
@@ -101,6 +101,7 @@ eEval (ix, line, ctx) = go where
             in OptionVal ty (Just eI)
     go AllField{} = StrLit tyStr line
     go (Field _ i) = StrLit tyStr (ctx ! (i-1)) -- cause vector indexing starts at 0
+    go LastField{} = StrLit tyStr (V.last ctx)
     go (EApp _ (UBuiltin _ IParse) e) =
         let eI = asStr (go e)
             in parseAsEInt eI
@@ -266,12 +267,12 @@ eEval (ix, line, ctx) = go where
         let str = asStr (go e)
             re = asRegex (go e')
             bss = splitBy re str
-            in Arr undefined (StrLit undefined <$> bss)
+            in Arr undefined (mkStr <$> bss)
     go (EApp _ (EApp _ (BBuiltin _ Splitc) e) e') =
         let str = asStr (go e)
             c = the (asStr (go e'))
             bss = BS.split c str
-            in Arr undefined (StrLit undefined <$> V.fromList bss)
+            in Arr undefined (mkStr <$> V.fromList bss)
     go (EApp _ (EApp _ (EApp _ (TBuiltin _ Substr) e0) e1) e2) =
         let eI0 = asStr (go e0)
             eI1 = asInt (go e1)
@@ -351,6 +352,9 @@ eEval (ix, line, ctx) = go where
     go (Cond _ p e0 e1) =
         let p' = asBool (go p)
             in if p' then go e0 else go e1
+    go (EApp _ (UBuiltin _ TallyList) e) =
+        let xs = asArr (go e)
+            in mkI $ fromIntegral $ V.length xs
     go e = error ("Internal error: " ++ show e)
 
 -- just shove some big number into the renamer and hope it doesn't clash (bad,
@@ -390,6 +394,8 @@ ir _ AllColumn{} = fmap mkStr
 ir re (Column _ i) = fmap (mkStr . atField re i)
 ir re (IParseCol _ i) = fmap (parseAsEInt . atField re i)
 ir re (FParseCol _ i) = fmap (parseAsF . atField re i)
+ir re (ParseCol ty@(TyApp _ _ (TyB _ TyFloat)) i) = ir re (FParseCol ty i)
+ir re (ParseCol ty@(TyApp _ _ (TyB _ TyInteger)) i) = ir re (IParseCol ty i)
 ir re (Implicit _ e) =
     imap (\ix line -> eEval (mkCtx re ix line) e)
 ir re (Guarded _ pe e) =
@@ -541,14 +547,15 @@ fileProcessor :: RurePtr
 fileProcessor _ AllField{}    = Left NakedField
 fileProcessor _ Field{}       = Left NakedField
 fileProcessor _ (NBuiltin _ Ix) = Left NakedField
-fileProcessor _ AllColumn{} = Right $ \inp ->
-    printStream $ fmap mkStr inp
-fileProcessor re (Column _ i) = Right $ \inp -> do
-    printStream $ fmap (mkStr . atField re i) inp
-fileProcessor re (IParseCol _ i) = Right $ \inp -> do
-    printStream $ fmap (parseAsEInt . atField re i) inp
-fileProcessor re (FParseCol _ i) = Right $ \inp -> do
-    printStream $ fmap (parseAsF . atField re i) inp
+fileProcessor re e@AllColumn{} = Right $ \inp ->
+    printStream $ ir re e inp
+fileProcessor re e@Column{} = Right $ \inp ->
+    printStream $ ir re e inp
+fileProcessor re e@IParseCol{} = Right $ \inp ->
+    printStream $ ir re e inp
+fileProcessor re e@FParseCol{} = Right $ \inp ->
+    printStream $ ir re e inp
+fileProcessor re e@ParseCol{} = Right $ \inp -> printStream $ ir re e inp
 fileProcessor re e@Guarded{} = Right $ \inp ->
     printStream $ ir re e inp
 fileProcessor re e@Implicit{} = Right $ \inp ->