+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.1.0.0
+
+  * Add builtin for last field
+  * Performance improvements
+  * Fix bug in how fields were split
+  * `:` works on column literals
+  * Add `#*` (list length) builtin
+
 # 1.0.0.0
 
   * Generalize type of `\.` (prior)